Project4_chat_app/requirements.txt.txt
cryptography==38.0.4
Pillow==9.5.0
pystray==0.19.3